+++ Engone/src/Engone/Components/Entity.cpp
@@ -6,7 +6,7 @@ namespace engone {
 	Entity::Entity(ComponentMask mask) : ComponentMask(mask) {
 		
 	}
-	Entity::Entity(ComponentMask mask, int* sizes, char* ptr) : ComponentMask(mask), componentSizes(sizes), stackPtr(ptr) {
+	Entity::Entity(ComponentMask mask, size_t* sizes, char* ptr) : ComponentMask(mask), componentSizes(sizes), stackPtr(ptr) {
 		
 	}
 }
